# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UEY Premium Mobility API",
        "description": "Complete backend API documentation for UEY Premium Mobility ride-hailing platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Default API Gateway"
        }
    ],
    "paths": {
...
        "/rides/estimate": {
            "post": {
                "tags": [
                    "Ride Booking"
                ],
                "summary": "Estimate Fare",
                "description": "Get fare estimates for all active vehicle types between pickup and destination.",
                "operationId": "a3eb936892224415e6525b099bb7fde5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstimateRideRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fare estimates retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "estimates": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "vehicle_type_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Standard"
                                                    },
                                                    "capacity": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "estimated_distance": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 2.34
                                                    },
                                                    "estimated_duration": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "estimated_fare": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 8.5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rides/request": {
            "post": {
                "tags": [
                    "Ride Booking"
                ],
                "summary": "Request Ride",
                "description": "Create a new ride request and find nearby drivers.",
                "operationId": "5a3bb80f73c0bcb61b2c1ddf93c99251",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestRideRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ride requested successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ride requested successfully."
                                        },
                                        "ride": {
                                            "$ref": "#/components/schemas/Ride"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rides/{ride}/cancel": {
            "post": {
                "tags": [
                    "Ride Booking"
                ],
                "summary": "Cancel Ride",
                "description": "Cancel an active ride request. Allowed only before trip starts (statuses: pending, accepted, arriving, arrived).",
                "operationId": "193b00dbacbeade2bfbd26d0b008c58d",
                "parameters": [
                    {
                        "name": "ride",
                        "in": "path",
                        "description": "ID of the Ride",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CancelRideRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ride cancelled successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ride cancelled successfully."
                                        },
                                        "ride": {
                                            "$ref": "#/components/schemas/Ride"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rides/{ride}": {
            "get": {
                "tags": [
                    "Ride Booking"
                ],
                "summary": "Get Ride Details",
                "description": "Get details of a specific ride by ID.",
                "operationId": "0df9dc743bddfe5f3436b49b60ae47fd",
                "parameters": [
                    {
                        "name": "ride",
                        "in": "path",
                        "description": "ID of the Ride",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ride details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "ride": {
                                            "$ref": "#/components/schemas/Ride"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "description": "Ride not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rides": {
            "get": {
                "tags": [
                    "Ride Booking"
                ],
                "summary": "Rider Ride History",
                "description": "Retrieve a list of past and active rides requested by the authenticated rider.",
                "operationId": "cd4dd887dd8a8cbdc176651994266f7e",
                "responses": {
                    "200": {
                        "description": "Ride history retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "rides": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ride"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rides/active": {
            "get": {
                "tags": [
                    "Ride Booking"
                ],
                "summary": "Get Active Ride",
                "description": "Get the current active (non-completed and non-cancelled) ride for the rider.",
                "operationId": "ec4d34d92bfa2746633fb8332a291765",
                "responses": {
                    "200": {
                        "description": "Active ride details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "ride": {
                                            "$ref": "#/components/schemas/Ride"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "description": "No active ride found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No active ride found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "EstimateRideRequest": {
                "required": [
                    "pickup_latitude",
                    "pickup_longitude",
                    "destination_latitude",
                    "destination_longitude"
                ],
                "properties": {
                    "pickup_latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 51.5074
                    },
                    "pickup_longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -0.1278
                    },
                    "destination_latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 51.5204
                    },
                    "destination_longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -0.1482
                    }
                },
                "type": "object"
            },
            "RequestRideRequest": {
                "required": [
                    "pickup_latitude",
                    "pickup_longitude",
                    "pickup_address",
                    "destination_latitude",
                    "destination_longitude",
                    "destination_address",
                    "vehicle_type_id"
                ],
                "properties": {
                    "pickup_latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 51.5074
                    },
                    "pickup_longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -0.1278
                    },
                    "pickup_address": {
                        "type": "string",
                        "example": "London Eye, London"
                    },
                    "destination_latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 51.5204
                    },
                    "destination_longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -0.1482
                    },
                    "destination_address": {
                        "type": "string",
                        "example": "Regent's Park, London"
                    },
                    "vehicle_type_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "CancelRideRequest": {
                "properties": {
                    "cancel_reason": {
                        "type": "string",
                        "example": "Plans changed"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "ValidationErrorResponse": {
                "description": "Validation errors in request parameters.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "The given data was invalid."
                                },
                                "errors": {
                                    "properties": {
                                        "phone": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The phone field is required."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UnauthorizedResponse": {
                "description": "Bearer token is missing, invalid, or expired.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Sanctum Bearer token for authorized requests (e.g. Bearer 1|...)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}